requirements.txt
click==8.1.3
flask==2.2.3
gunicorn==20.1
itsdangerous==2.1.2
jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
requests==2.28.2
beautifulsoup4==4.11.2
statistics==1.0.3.5
DateTime==5.0
discord-webhook==1.0.0
python-dotenv==0.21.1
openai==0.26.5